Track metadata in tree snapshot tests
